% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dog_breed_dataset}
\alias{dog_breed_dataset}
\title{Dog breed dataset}
\usage{
dog_breed_dataset(
  root,
  split = "train",
  transform = NULL,
  download = FALSE,
  loader = torchvision::base_loader,
  force = FALSE
)
}
\arguments{
\item{root}{path to the data location.}

\item{split}{string. 'train' or 'submission'.}

\item{transform}{function that receives a torch tensor and return another torch tensor, transformed.}

\item{download}{wether to download or not.}

\item{loader}{function. first argument is path for image. See \link[torchvision:base_loader]{torchvision::base_loader}.}

\item{force}{force download?}
}
\description{
Prepares the Dog Breed dataset available in Kaggle \href{https://www.kaggle.com/c/dog-breed-identification}{here}
}
\details{
We use Kaggle API CLI to download the dataset.If you want to download the dataset.
}
