% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nn_dog}
\alias{nn_dog}
\title{DogNet}
\usage{
nn_dog(classes = 100, freeze = TRUE, device = "cpu")
}
\arguments{
\item{classes}{(int) number of output probabilities.}

\item{freeze}{(bool) should resnet18 weights be frozen?}

\item{device}{(chr or torch_device)}
}
\description{
nn_module that is just a resnet18 with a custom endpoint with 'classes' output.
}
